instructions for disabling persistent rfkill on headless pi
If an operator fails to set the country code in `wpa_supplicant.conf` when
configuring a headless pi, the `wifi-county.service` unit will
persistently disable the wifi interface.  Without an alternate means
of ingress (serial console or ethernet), the operator will be unable
to access the Pi.

This commit updates the documentation with the necessary steps to
recover from this situation.

This addresses the situation discussed in #raspberrypi-sys-mods/27.